fix: resolve all GitHub CodeQL security/quality alerts
Fix 9 integer-multiplication-cast-to-long warnings across 6 files:
- wmo_renderer.cpp: grid cell count and height variance calculation
- composite_renderer.cpp: overlay tile grid allocation
- vk_texture.cpp: image size calculation (width*height*bpp)
- m2_renderer.cpp: collision grid cell allocation
- character_renderer.cpp: normal map buffer and height variance
- world_entry_callback_handler.cpp: tile reserve count

All fixes cast operands to size_t/double before multiplication to
prevent integer overflow when dimensions are large.

+++ src/core/world_entry_callback_handler.cpp
@@ -108,7 +108,7 @@ static void precacheNearbyTiles(rendering::TerrainManager* terrainMgr,
     auto [tileX, tileY] = core::coords::worldToTile(renderPos.x, renderPos.y);
     int side = 2 * radius + 1;
     std::vector<std::pair<int,int>> tiles;
-    tiles.reserve(side * side);
+    tiles.reserve(static_cast<size_t>(side) * static_cast<size_t>(side));
     for (int dy = -radius; dy <= radius; dy++)
         for (int dx = -radius; dx <= radius; dx++)
             tiles.push_back({tileX + dx, tileY + dy});
